Restructure project
- Make every CLI command group a seperate component, by giving it a
  directory directly below `turret/`.
- Assign every component to its own CLI interface. The original
  turret.cli file is split up over multiple <component>.cli files.
- Move Nmap related function to a module in turret.raw
